requirements.txt
PyQt6
google-genai
protobuf
google-api-core
google-auth
Pygments
markdown-it-py
psutil  
httpx  
pydantic
pylint
coverage